ingestion/requirements.txt
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
textblob==0.17.1
nltk==3.8.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
yfinance==0.2.28
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0 